refactor: share the global composer paste listener between apps
