+++ apps/ui/src/ui-classic/components/session-view/composer/index.tsx
@@ -9,7 +9,7 @@ import {
 	watchComposerAttachmentTextScroll,
 	type ComposerAttachmentHoverPreviewState,
 } from '@studio/common/ai/composer-attachment-preview';
-import { getComposerClipboardFiles } from '@studio/common/ai/composer-attachments';
+import { watchComposerFilePaste } from '@studio/common/ai/composer-attachments';
 import { AI_MODELS, getAiModelFamily, getAiModelLabel } from '@studio/common/ai/models';
 import { isStudioCustomEntryOfType } from '@studio/common/ai/sessions/entry-types';
 import { AI_SKILL_COMMANDS } from '@studio/common/ai/slash-commands';
@@ -423,30 +423,11 @@ export const Composer = forwardRef< ComposerHandle, ComposerProps >( function Co
 		}
 	}, [ autoFocus, sessionId ] );
 
-	// Accept file pastes anywhere in the view (e.g. after clicking the
-	// conversation), like other AI chat tools. The textarea's own onPaste
-	// prevents default first, so `defaultPrevented` avoids double-adding;
-	// pastes inside open dialogs are left alone.
 	useEffect( () => {
-		const onDocumentPaste = ( event: ClipboardEvent ) => {
-			if ( event.defaultPrevented || ! event.clipboardData ) {
-				return;
-			}
-			if ( event.target instanceof Element && event.target.closest( '[role="dialog"]' ) ) {
-				return;
-			}
-			const files = getComposerClipboardFiles( event.clipboardData );
-			if ( files.length === 0 ) {
-				return;
-			}
-			event.preventDefault();
+		return watchComposerFilePaste( ( files ) => {
 			void addFiles( files );
 			textareaRef.current?.focus();
-		};
-		document.addEventListener( 'paste', onDocumentPaste );
-		return () => {
-			document.removeEventListener( 'paste', onDocumentPaste );
-		};
+		} );
 	}, [ addFiles ] );
 
 	useLayoutEffect( () => {
