+++ packages/common/ai/composer-attachments.ts
@@ -227,6 +227,34 @@ export function getComposerClipboardFiles( dataTransfer: DataTransfer ): File[] 
 		.map( normalizePastedFile );
 }
 
+/**
+ * Listen for file pastes anywhere in the document (e.g. with focus on the
+ * conversation), like other AI chat tools, and hand the clipboard files to
+ * `onFiles`. A composer textarea's own paste handler prevents default first,
+ * so `defaultPrevented` avoids double-adding; pastes inside open dialogs are
+ * left alone. Returns an unsubscribe function.
+ */
+export function watchComposerFilePaste( onFiles: ( files: File[] ) => void ): () => void {
+	const onPaste = ( event: ClipboardEvent ) => {
+		if ( event.defaultPrevented || ! event.clipboardData ) {
+			return;
+		}
+		if ( event.target instanceof Element && event.target.closest( '[role="dialog"]' ) ) {
+			return;
+		}
+		const files = getComposerClipboardFiles( event.clipboardData );
+		if ( files.length === 0 ) {
+			return;
+		}
+		event.preventDefault();
+		onFiles( files );
+	};
+	document.addEventListener( 'paste', onPaste );
+	return () => {
+		document.removeEventListener( 'paste', onPaste );
+	};
+}
+
 function addUniqueError( errors: string[], error: string ): void {
 	if ( ! errors.includes( error ) ) {
 		errors.push( error );
